added tests for Image, Album, and ImageSync

+++ app/Album.php
@@ -17,45 +17,71 @@ class Album extends Model {
     }
 
     /**
-     * Adds an image to the album.
+     * Adds an image to the album using am Image or an id.
      *
-     * @param Image $image
+     * @param Image | integer $image
      */
     public function addImage($image)
     {
-        return $this->images()->attach($image);
+        if ($image instanceof Image)
+            $image = $image->id;
+
+        if (is_numeric($image))
+            return $this->images()->attach($image);
     }
 
     /**
-     * Removes an image from the album.
+     * Removes an image from the album using am Image or an id.
      *
-     * @param $image
+     * @param Image | integer $image
      * @return int
      */
     public function removeImage($image)
     {
-        return $this->images()->detach($image);
+        if ($image instanceof Image)
+            $image = $image->id;
+
+        if (is_numeric($image))
+            return $this->images()->detach($image);
     }
 
     /**
-     * Adds each image in the array to the album.
+     * Adds each image in the array to the album. Array may contain
+     * Images or id numbers.
      *
-     * @param array $images
+     * @param Image[] | integer[] | mixed[] $images
      */
     public function addImages($images)
     {
-        if (is_array($images))
-            array_map('addImage', $images);
+        if (is_array($images)) {
+            array_map(function($image) {
+                $this->addImage($image);
+            }, $images);
+        }
     }
 
     /**
-     * Removes each image in the array from the album.
+     * Removes each image in the array from the album. Array may
+     * contain Images or id numbers.
      *
-     * @param array $images
+     * @param Image[] | integer[] | mixed[] $images
      */
     public function removeImages($images)
     {
-        if(is_array($images))
-            array_map('removeImage', $images);
+        if(is_array($images)) {
+            array_map(function($image) {
+                $this->removeImage($image);
+            }, $images);
+        }
+    }
+
+    /**
+     * Removes all of the images from the album.
+     *
+     * @return int
+     */
+    public function removeAll()
+    {
+        return $this->images()->detach();
     }
 }
